Alt-Text Generator/backend/requirements.txt
flask
flask-cors
transformers
torch
Pillow
